requirements.txt
certifi
dill
fastapi
imblearn
pandas
pymongo
PyYAML
scikit-learn
scipy
uvicorn
xgboost
-e .